pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raspberry-translator"
version = "1.0.0"
description = "Real-time audio translator for Raspberry Pi 5"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "silero-vad>=5.0",
    "scipy>=1.10.0",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20.0",
    "huggingface-hub>=0.16.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
    "httpx>=0.24",
]
web = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "httpx>=0.24",
]

[project.scripts]
raspberry-translator = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
